requirements.txt
asgiref==3.5.0
asttokens==2.0.5
backcall==0.2.0
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangorestframework-xml==2.0.0
drf-yasg==1.20.0
executing==0.8.3
httpie==3.1.0
idna==3.3
inflection==0.5.1
ipython==8.2.0
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.8
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
responses==0.20.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
traitlets==5.1.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
